細かい修正、テスト追加に伴うバグ修正 (#146)
* playlist, thumbnailをちゃんとPOSTするように

* 枠数調整、サムネイルやプレイリストのURLをPOSTできるように

* update reservation slot

* logging

* add test

* add test

* add logging

* add logging

* fix bug

* limit livecomments

* limit reactions

* お気に入り絵文字

* テスト追加

* add tests & fix endpoint

+++ bench/isupipe/client_initialize.go
@@ -8,7 +8,7 @@ import (
 )
 
 type InitializeResponse struct {
-	AdvertiseLevel int    `json:"advertise_level"`
+	AdvertiseLevel int64  `json:"advertise_level"`
 	Language       string `json:"language"`
 }
